Remove CacheItemPriority and add sample eviction strategies
