--- src/Microsoft.Extensions.Caching.Abstractions/CacheItemPriority.cs
@@ -1,17 +1,0 @@
-// Copyright (c) .NET Foundation. All rights reserved.
-// Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
-
-namespace Microsoft.Extensions.Caching.Memory
-{
-    // TODO: Granularity?
-    /// <summary>
-    /// Specifies how items are prioritized for preservation during a memory pressure triggered cleanup.
-    /// </summary>
-    public enum CacheItemPriority
-    {
-        Low,
-        Normal,
-        High,
-        NeverRemove,
-    }
-}
