+++ src/Microsoft.Extensions.Caching.Abstractions/ICacheEntry.cs
@@ -48,11 +48,5 @@ namespace Microsoft.Extensions.Caching.Memory
         /// </summary>
         IList<PostEvictionCallbackRegistration> PostEvictionCallbacks { get; }
 
-        /// <summary>
-        /// Gets or sets the priority for keeping the cache entry in the cache during a
-        /// memory pressure triggered cleanup. The default is <see cref="CacheItemPriority.Normal"/>.
-        /// </summary>
-        CacheItemPriority Priority { get; set; }
-
     }
 }
