+++ src/Microsoft.Extensions.Caching.Abstractions/MemoryCacheEntryOptions.cs
@@ -84,11 +84,5 @@ namespace Microsoft.Extensions.Caching.Memory
         /// </summary>
         public IList<PostEvictionCallbackRegistration> PostEvictionCallbacks { get; }
             = new List<PostEvictionCallbackRegistration>();
-
-        /// <summary>
-        /// Gets or sets the priority for keeping the cache entry in the cache during a
-        /// memory pressure triggered cleanup. The default is <see cref="CacheItemPriority.Normal"/>.
-        /// </summary>
-        public CacheItemPriority Priority { get; set; } = CacheItemPriority.Normal;
     }
 }
